refactor: rename "cache::requests" mod to "cache::messages"
In Rust, we do not decouple requests from responses, unlike in other
SDKs. Because of this, the requests and responses are in the same
modules. For clarity we rename the `requests` mod to `messages`. This
better captures this.
