+++ src/cache/messages/dictionary/dictionary_set_field.rs
@@ -1,4 +1,4 @@
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::MomentoRequest;
 use crate::cache::CollectionTtl;
 use crate::utils::prep_request_with_timeout;
 use crate::IntoBytes;
