+++ src/cache/messages/list/list_pop_front.rs
