+++ src/cache/messages/scalar/delete.rs
