+++ src/cache/messages/scalar/get.rs
@@ -1,4 +1,4 @@
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::MomentoRequest;
 use crate::utils::parse_string;
 use crate::utils::prep_request_with_timeout;
 use crate::CacheClient;
