+++ src/cache/messages/scalar/key_exists.rs
