+++ src/cache/messages/scalar/keys_exist.rs
