+++ src/cache/messages/scalar/set_if_equal.rs
@@ -1,21 +1,21 @@
-use momento_protos::cache_client::set_if_request::Condition::NotEqual;
+use momento_protos::cache_client::set_if_request::Condition::Equal;
 use momento_protos::cache_client::set_if_response;
 
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::MomentoRequest;
 use crate::utils::prep_request_with_timeout;
 use crate::CacheClient;
 use crate::{IntoBytes, MomentoError, MomentoResult};
 use std::time::Duration;
 
-/// Request to associate the given key with the given value if the key does not already exist
-/// in the cache or the value in the cache is not equal to the value supplied `not_equal` value.
+/// Request to associate the given key with the given value if the key is present
+/// in the cache and its value is equal to the supplied `equal` value.
 ///
 /// # Arguments
 ///
 /// * `cache_name` - The name of the cache to create.
 /// * `key` - key of the item whose value we are setting
 /// * `value` - data to store
-/// * `not_equal` - data to compare to the cached value
+/// * `equal` - data to compare to the cached value
 ///
 /// # Optional Arguments
 ///
@@ -28,11 +28,11 @@ use std::time::Duration;
 /// # use momento_test_util::create_doctest_cache_client;
 /// # tokio_test::block_on(async {
 /// use std::time::Duration;
-/// use momento::cache::{SetIfNotEqual, SetIfNotEqualRequest};
+/// use momento::cache::{SetIfEqual, SetIfEqualRequest};
 /// use momento::MomentoErrorCode;
 /// # let (cache_client, cache_name) = create_doctest_cache_client();
 ///
-/// let set_request = SetIfNotEqualRequest::new(
+/// let set_request = SetIfEqualRequest::new(
 ///     &cache_name,
 ///     "key",
 ///     "new-value",
@@ -41,8 +41,8 @@ use std::time::Duration;
 ///
 /// match cache_client.send_request(set_request).await {
 ///     Ok(response) => match response {
-///         SetIfNotEqual::Stored => println!("Value stored"),
-///         SetIfNotEqual::NotStored => println!("Value not stored"),
+///         SetIfEqual::Stored => println!("Value stored"),
+///         SetIfEqual::NotStored => println!("Value not stored"),
 ///     }
 ///     Err(e) => if let MomentoErrorCode::NotFoundError = e.error_code {
 ///         println!("Cache not found: {}", &cache_name);
@@ -54,7 +54,7 @@ use std::time::Duration;
 /// # })
 /// # }
 /// ```
-pub struct SetIfNotEqualRequest<K: IntoBytes, V: IntoBytes, E: IntoBytes> {
+pub struct SetIfEqualRequest<K: IntoBytes, V: IntoBytes, E: IntoBytes> {
     cache_name: String,
     key: K,
     value: V,
@@ -62,7 +62,7 @@ pub struct SetIfNotEqualRequest<K: IntoBytes, V: IntoBytes, E: IntoBytes> {
     ttl: Option<Duration>,
 }
 
-impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> SetIfNotEqualRequest<K, V, E> {
+impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> SetIfEqualRequest<K, V, E> {
     pub fn new(cache_name: impl Into<String>, key: K, value: V, equal: E) -> Self {
         let ttl = None;
         Self {
@@ -80,10 +80,10 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> SetIfNotEqualRequest<K, V, E> {
     }
 }
 
-impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest for SetIfNotEqualRequest<K, V, E> {
-    type Response = SetIfNotEqual;
+impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest for SetIfEqualRequest<K, V, E> {
+    type Response = SetIfEqual;
 
-    async fn send(self, cache_client: &CacheClient) -> MomentoResult<SetIfNotEqual> {
+    async fn send(self, cache_client: &CacheClient) -> MomentoResult<SetIfEqual> {
         let request = prep_request_with_timeout(
             &self.cache_name,
             cache_client.configuration.deadline_millis(),
@@ -91,7 +91,7 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest for SetIfNotEqualR
                 cache_key: self.key.into_bytes(),
                 cache_body: self.value.into_bytes(),
                 ttl_milliseconds: cache_client.expand_ttl_ms(self.ttl)?,
-                condition: Some(NotEqual(momento_protos::common::NotEqual {
+                condition: Some(Equal(momento_protos::common::Equal {
                     value_to_check: self.equal.into_bytes(),
                 })),
             },
@@ -104,10 +104,10 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest for SetIfNotEqualR
             .await?
             .into_inner();
         match response.result {
-            Some(set_if_response::Result::Stored(_)) => Ok(SetIfNotEqual::Stored),
-            Some(set_if_response::Result::NotStored(_)) => Ok(SetIfNotEqual::NotStored),
+            Some(set_if_response::Result::Stored(_)) => Ok(SetIfEqual::Stored),
+            Some(set_if_response::Result::NotStored(_)) => Ok(SetIfEqual::NotStored),
             _ => Err(MomentoError::unknown_error(
-                "SetIfNotEqual",
+                "SetIfEqual",
                 Some(format!("{:#?}", response)),
             )),
         }
@@ -115,7 +115,7 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest for SetIfNotEqualR
 }
 
 #[derive(Debug, PartialEq, Eq)]
-pub enum SetIfNotEqual {
+pub enum SetIfEqual {
     Stored,
     NotStored,
 }
