+++ src/cache/messages/scalar/set_if_not_equal.rs
@@ -1,19 +1,21 @@
-use momento_protos::cache_client::set_if_request::Condition::Absent;
+use momento_protos::cache_client::set_if_request::Condition::NotEqual;
 use momento_protos::cache_client::set_if_response;
 
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::MomentoRequest;
 use crate::utils::prep_request_with_timeout;
 use crate::CacheClient;
 use crate::{IntoBytes, MomentoError, MomentoResult};
 use std::time::Duration;
 
-/// Request to set associate the given key with the given value if key is not already present in the cache.
+/// Request to associate the given key with the given value if the key does not already exist
+/// in the cache or the value in the cache is not equal to the value supplied `not_equal` value.
 ///
 /// # Arguments
 ///
 /// * `cache_name` - The name of the cache to create.
 /// * `key` - key of the item whose value we are setting
 /// * `value` - data to store
+/// * `not_equal` - data to compare to the cached value
 ///
 /// # Optional Arguments
 ///
@@ -26,20 +28,21 @@ use std::time::Duration;
 /// # use momento_test_util::create_doctest_cache_client;
 /// # tokio_test::block_on(async {
 /// use std::time::Duration;
-/// use momento::cache::{SetIfAbsent, SetIfAbsentRequest};
+/// use momento::cache::{SetIfNotEqual, SetIfNotEqualRequest};
 /// use momento::MomentoErrorCode;
 /// # let (cache_client, cache_name) = create_doctest_cache_client();
 ///
-/// let set_request = SetIfAbsentRequest::new(
+/// let set_request = SetIfNotEqualRequest::new(
 ///     &cache_name,
 ///     "key",
-///     "value1"
+///     "new-value",
+///     "cached-value"
 /// ).ttl(Duration::from_secs(60));
 ///
 /// match cache_client.send_request(set_request).await {
 ///     Ok(response) => match response {
-///         SetIfAbsent::Stored => println!("Value stored"),
-///         SetIfAbsent::NotStored => println!("Value not stored"),
+///         SetIfNotEqual::Stored => println!("Value stored"),
+///         SetIfNotEqual::NotStored => println!("Value not stored"),
 ///     }
 ///     Err(e) => if let MomentoErrorCode::NotFoundError = e.error_code {
 ///         println!("Cache not found: {}", &cache_name);
@@ -51,20 +54,22 @@ use std::time::Duration;
 /// # })
 /// # }
 /// ```
-pub struct SetIfAbsentRequest<K: IntoBytes, V: IntoBytes> {
+pub struct SetIfNotEqualRequest<K: IntoBytes, V: IntoBytes, E: IntoBytes> {
     cache_name: String,
     key: K,
     value: V,
+    equal: E,
     ttl: Option<Duration>,
 }
 
-impl<K: IntoBytes, V: IntoBytes> SetIfAbsentRequest<K, V> {
-    pub fn new(cache_name: impl Into<String>, key: K, value: V) -> Self {
+impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> SetIfNotEqualRequest<K, V, E> {
+    pub fn new(cache_name: impl Into<String>, key: K, value: V, equal: E) -> Self {
         let ttl = None;
         Self {
             cache_name: cache_name.into(),
             key,
             value,
+            equal,
             ttl,
         }
     }
@@ -75,10 +80,10 @@ impl<K: IntoBytes, V: IntoBytes> SetIfAbsentRequest<K, V> {
     }
 }
 
-impl<K: IntoBytes, V: IntoBytes> MomentoRequest for SetIfAbsentRequest<K, V> {
-    type Response = SetIfAbsent;
+impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest for SetIfNotEqualRequest<K, V, E> {
+    type Response = SetIfNotEqual;
 
-    async fn send(self, cache_client: &CacheClient) -> MomentoResult<SetIfAbsent> {
+    async fn send(self, cache_client: &CacheClient) -> MomentoResult<SetIfNotEqual> {
         let request = prep_request_with_timeout(
             &self.cache_name,
             cache_client.configuration.deadline_millis(),
@@ -86,7 +91,9 @@ impl<K: IntoBytes, V: IntoBytes> MomentoRequest for SetIfAbsentRequest<K, V> {
                 cache_key: self.key.into_bytes(),
                 cache_body: self.value.into_bytes(),
                 ttl_milliseconds: cache_client.expand_ttl_ms(self.ttl)?,
-                condition: Some(Absent(momento_protos::common::Absent {})),
+                condition: Some(NotEqual(momento_protos::common::NotEqual {
+                    value_to_check: self.equal.into_bytes(),
+                })),
             },
         )?;
 
@@ -97,10 +104,10 @@ impl<K: IntoBytes, V: IntoBytes> MomentoRequest for SetIfAbsentRequest<K, V> {
             .await?
             .into_inner();
         match response.result {
-            Some(set_if_response::Result::Stored(_)) => Ok(SetIfAbsent::Stored),
-            Some(set_if_response::Result::NotStored(_)) => Ok(SetIfAbsent::NotStored),
+            Some(set_if_response::Result::Stored(_)) => Ok(SetIfNotEqual::Stored),
+            Some(set_if_response::Result::NotStored(_)) => Ok(SetIfNotEqual::NotStored),
             _ => Err(MomentoError::unknown_error(
-                "SetIfAbsent",
+                "SetIfNotEqual",
                 Some(format!("{:#?}", response)),
             )),
         }
@@ -108,7 +115,7 @@ impl<K: IntoBytes, V: IntoBytes> MomentoRequest for SetIfAbsentRequest<K, V> {
 }
 
 #[derive(Debug, PartialEq, Eq)]
-pub enum SetIfAbsent {
+pub enum SetIfNotEqual {
     Stored,
     NotStored,
 }
