+++ src/cache/messages/scalar/set_if_present.rs
@@ -1,21 +1,19 @@
-use momento_protos::cache_client::set_if_request::Condition::AbsentOrEqual;
+use momento_protos::cache_client::set_if_request::Condition::Present;
 use momento_protos::cache_client::set_if_response;
 
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::MomentoRequest;
 use crate::utils::prep_request_with_timeout;
 use crate::CacheClient;
 use crate::{IntoBytes, MomentoError, MomentoResult};
 use std::time::Duration;
 
-/// Request to associate the given key with the given value if the key does not already
-/// exist in the cache or the value in the cache is equal to the supplied `equal` value.
+/// Request to set associate the given key with the given value if key is present in the cache.
 ///
 /// # Arguments
 ///
 /// * `cache_name` - The name of the cache to create.
 /// * `key` - key of the item whose value we are setting
 /// * `value` - data to store
-/// * `equal` - data to compare to the cached value
 ///
 /// # Optional Arguments
 ///
@@ -28,21 +26,20 @@ use std::time::Duration;
 /// # use momento_test_util::create_doctest_cache_client;
 /// # tokio_test::block_on(async {
 /// use std::time::Duration;
-/// use momento::cache::{SetIfAbsentOrEqual, SetIfAbsentOrEqualRequest};
+/// use momento::cache::{SetIfPresent, SetIfPresentRequest};
 /// use momento::MomentoErrorCode;
 /// # let (cache_client, cache_name) = create_doctest_cache_client();
 ///
-/// let set_request = SetIfAbsentOrEqualRequest::new(
+/// let set_request = SetIfPresentRequest::new(
 ///     &cache_name,
 ///     "key",
-///     "new-value",
-///     "cached-value"
+///     "value1"
 /// ).ttl(Duration::from_secs(60));
 ///
 /// match cache_client.send_request(set_request).await {
 ///     Ok(response) => match response {
-///         SetIfAbsentOrEqual::Stored => println!("Value stored"),
-///         SetIfAbsentOrEqual::NotStored => println!("Value not stored"),
+///         SetIfPresent::Stored => println!("Value stored"),
+///         SetIfPresent::NotStored => println!("Value not stored"),
 ///     }
 ///     Err(e) => if let MomentoErrorCode::NotFoundError = e.error_code {
 ///         println!("Cache not found: {}", &cache_name);
@@ -54,22 +51,20 @@ use std::time::Duration;
 /// # })
 /// # }
 /// ```
-pub struct SetIfAbsentOrEqualRequest<K: IntoBytes, V: IntoBytes, E: IntoBytes> {
+pub struct SetIfPresentRequest<K: IntoBytes, V: IntoBytes> {
     cache_name: String,
     key: K,
     value: V,
-    equal: E,
     ttl: Option<Duration>,
 }
 
-impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> SetIfAbsentOrEqualRequest<K, V, E> {
-    pub fn new(cache_name: impl Into<String>, key: K, value: V, equal: E) -> Self {
+impl<K: IntoBytes, V: IntoBytes> SetIfPresentRequest<K, V> {
+    pub fn new(cache_name: impl Into<String>, key: K, value: V) -> Self {
         let ttl = None;
         Self {
             cache_name: cache_name.into(),
             key,
             value,
-            equal,
             ttl,
         }
     }
@@ -80,12 +75,10 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> SetIfAbsentOrEqualRequest<K, V, E
     }
 }
 
-impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest
-    for SetIfAbsentOrEqualRequest<K, V, E>
-{
-    type Response = SetIfAbsentOrEqual;
+impl<K: IntoBytes, V: IntoBytes> MomentoRequest for SetIfPresentRequest<K, V> {
+    type Response = SetIfPresent;
 
-    async fn send(self, cache_client: &CacheClient) -> MomentoResult<SetIfAbsentOrEqual> {
+    async fn send(self, cache_client: &CacheClient) -> MomentoResult<SetIfPresent> {
         let request = prep_request_with_timeout(
             &self.cache_name,
             cache_client.configuration.deadline_millis(),
@@ -93,9 +86,7 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest
                 cache_key: self.key.into_bytes(),
                 cache_body: self.value.into_bytes(),
                 ttl_milliseconds: cache_client.expand_ttl_ms(self.ttl)?,
-                condition: Some(AbsentOrEqual(momento_protos::common::AbsentOrEqual {
-                    value_to_check: self.equal.into_bytes(),
-                })),
+                condition: Some(Present(momento_protos::common::Present {})),
             },
         )?;
 
@@ -106,10 +97,10 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest
             .await?
             .into_inner();
         match response.result {
-            Some(set_if_response::Result::Stored(_)) => Ok(SetIfAbsentOrEqual::Stored),
-            Some(set_if_response::Result::NotStored(_)) => Ok(SetIfAbsentOrEqual::NotStored),
+            Some(set_if_response::Result::Stored(_)) => Ok(SetIfPresent::Stored),
+            Some(set_if_response::Result::NotStored(_)) => Ok(SetIfPresent::NotStored),
             _ => Err(MomentoError::unknown_error(
-                "SetIfAbsentOrEqual",
+                "SetIfPresent",
                 Some(format!("{:#?}", response)),
             )),
         }
@@ -117,7 +108,7 @@ impl<K: IntoBytes, V: IntoBytes, E: IntoBytes> MomentoRequest
 }
 
 #[derive(Debug, PartialEq, Eq)]
-pub enum SetIfAbsentOrEqual {
+pub enum SetIfPresent {
     Stored,
     NotStored,
 }
