+++ src/cache/messages/set/set_add_elements.rs
@@ -1,6 +1,6 @@
 use momento_protos::cache_client::SetUnionRequest;
 
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::MomentoRequest;
 use crate::cache::CollectionTtl;
 use crate::utils::prep_request_with_timeout;
 use crate::CacheClient;
