+++ src/cache/messages/sorted_set/sorted_set_fetch_by_score.rs
@@ -3,10 +3,10 @@ use momento_protos::cache_client::sorted_set_fetch_request::{by_score, ByScore, 
 use momento_protos::cache_client::SortedSetFetchRequest;
 use momento_protos::common::Unbounded;
 
-use crate::cache::requests::sorted_set::sorted_set_fetch_by_rank::SortedSetOrder;
-use crate::cache::requests::sorted_set::sorted_set_fetch_by_rank::SortedSetOrder::Ascending;
-use crate::cache::requests::sorted_set::sorted_set_fetch_response::SortedSetFetch;
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::sorted_set::sorted_set_fetch_by_rank::SortedSetOrder;
+use crate::cache::messages::sorted_set::sorted_set_fetch_by_rank::SortedSetOrder::Ascending;
+use crate::cache::messages::sorted_set::sorted_set_fetch_response::SortedSetFetch;
+use crate::cache::messages::MomentoRequest;
 use crate::utils::prep_request_with_timeout;
 use crate::{CacheClient, IntoBytes, MomentoResult};
