+++ src/cache/messages/sorted_set/sorted_set_put_element.rs
@@ -1,6 +1,6 @@
 use momento_protos::cache_client::{SortedSetElement, SortedSetPutRequest};
 
-use crate::cache::requests::MomentoRequest;
+use crate::cache::messages::MomentoRequest;
 use crate::cache::CollectionTtl;
 use crate::utils::prep_request_with_timeout;
 use crate::{CacheClient, IntoBytes, MomentoResult};
