+++ src/cache/mod.rs
@@ -1,94 +1,92 @@
-// Make module "requests" private and re-export its contents so users can
-// import using "momento::cache::{Get, Set, ...}"
-mod requests;
+mod messages;
 
-pub use requests::MomentoRequest;
+pub use messages::MomentoRequest;
 
-pub use requests::create_cache::{CreateCache, CreateCacheRequest};
-pub use requests::delete_cache::{DeleteCache, DeleteCacheRequest};
-pub use requests::flush_cache::{FlushCache, FlushCacheRequest};
-pub use requests::list_caches::{
+pub use messages::create_cache::{CreateCache, CreateCacheRequest};
+pub use messages::delete_cache::{DeleteCache, DeleteCacheRequest};
+pub use messages::flush_cache::{FlushCache, FlushCacheRequest};
+pub use messages::list_caches::{
     CacheInfo, CacheLimits, ListCaches, ListCachesRequest, TopicLimits,
 };
 
-pub use requests::dictionary::dictionary_fetch::{
+pub use messages::dictionary::dictionary_fetch::{
     DictionaryFetch, DictionaryFetchRequest, DictionaryFetchValue,
 };
-pub use requests::dictionary::dictionary_get_field::{
+pub use messages::dictionary::dictionary_get_field::{
     DictionaryGetField, DictionaryGetFieldRequest,
 };
-pub use requests::dictionary::dictionary_get_fields::{
+pub use messages::dictionary::dictionary_get_fields::{
     DictionaryGetFields, DictionaryGetFieldsRequest,
 };
-pub use requests::dictionary::dictionary_increment::{
+pub use messages::dictionary::dictionary_increment::{
     DictionaryIncrement, DictionaryIncrementRequest,
 };
-pub use requests::dictionary::dictionary_length::{DictionaryLength, DictionaryLengthRequest};
-pub use requests::dictionary::dictionary_remove_field::{
+pub use messages::dictionary::dictionary_length::{DictionaryLength, DictionaryLengthRequest};
+pub use messages::dictionary::dictionary_remove_field::{
     DictionaryRemoveField, DictionaryRemoveFieldRequest,
 };
-pub use requests::dictionary::dictionary_remove_fields::{
+pub use messages::dictionary::dictionary_remove_fields::{
     DictionaryRemoveFields, DictionaryRemoveFieldsRequest,
 };
-pub use requests::dictionary::dictionary_set_field::{
+pub use messages::dictionary::dictionary_set_field::{
     DictionarySetField, DictionarySetFieldRequest,
 };
-pub use requests::dictionary::dictionary_set_fields::{
+pub use messages::dictionary::dictionary_set_fields::{
     DictionaryFieldValuePair, DictionarySetFields, DictionarySetFieldsRequest,
     IntoDictionaryFieldValuePairs,
 };
 
-pub use requests::scalar::decrease_ttl::{DecreaseTtl, DecreaseTtlRequest};
-pub use requests::scalar::delete::{Delete, DeleteRequest};
-pub use requests::scalar::get::{Get, GetRequest, GetValue};
-pub use requests::scalar::increase_ttl::{IncreaseTtl, IncreaseTtlRequest};
-pub use requests::scalar::increment::{Increment, IncrementRequest};
-pub use requests::scalar::item_get_ttl::{ItemGetTtl, ItemGetTtlRequest};
-pub use requests::scalar::item_get_type::{ItemGetType, ItemGetTypeRequest, ItemType};
-pub use requests::scalar::key_exists::{KeyExists, KeyExistsRequest};
-pub use requests::scalar::keys_exist::{KeysExist, KeysExistRequest};
-pub use requests::scalar::set::{Set, SetRequest};
-pub use requests::scalar::set_if_absent::{SetIfAbsent, SetIfAbsentRequest};
-pub use requests::scalar::set_if_absent_or_equal::{SetIfAbsentOrEqual, SetIfAbsentOrEqualRequest};
-pub use requests::scalar::set_if_equal::{SetIfEqual, SetIfEqualRequest};
-pub use requests::scalar::set_if_not_equal::{SetIfNotEqual, SetIfNotEqualRequest};
-pub use requests::scalar::set_if_present::{SetIfPresent, SetIfPresentRequest};
-pub use requests::scalar::set_if_present_and_not_equal::{
+pub use messages::scalar::decrease_ttl::{DecreaseTtl, DecreaseTtlRequest};
+pub use messages::scalar::delete::{Delete, DeleteRequest};
+pub use messages::scalar::get::{Get, GetRequest, GetValue};
+pub use messages::scalar::increase_ttl::{IncreaseTtl, IncreaseTtlRequest};
+pub use messages::scalar::increment::{Increment, IncrementRequest};
+pub use messages::scalar::item_get_ttl::{ItemGetTtl, ItemGetTtlRequest};
+pub use messages::scalar::item_get_type::{ItemGetType, ItemGetTypeRequest, ItemType};
+pub use messages::scalar::key_exists::{KeyExists, KeyExistsRequest};
+pub use messages::scalar::keys_exist::{KeysExist, KeysExistRequest};
+pub use messages::scalar::set::{Set, SetRequest};
+pub use messages::scalar::set_if_absent::{SetIfAbsent, SetIfAbsentRequest};
+pub use messages::scalar::set_if_absent_or_equal::{SetIfAbsentOrEqual, SetIfAbsentOrEqualRequest};
+pub use messages::scalar::set_if_equal::{SetIfEqual, SetIfEqualRequest};
+pub use messages::scalar::set_if_not_equal::{SetIfNotEqual, SetIfNotEqualRequest};
+pub use messages::scalar::set_if_present::{SetIfPresent, SetIfPresentRequest};
+pub use messages::scalar::set_if_present_and_not_equal::{
     SetIfPresentAndNotEqual, SetIfPresentAndNotEqualRequest,
 };
-pub use requests::scalar::update_ttl::{UpdateTtl, UpdateTtlRequest};
+pub use messages::scalar::update_ttl::{UpdateTtl, UpdateTtlRequest};
 
-pub use requests::set::set_add_elements::{SetAddElements, SetAddElementsRequest};
-pub use requests::set::set_fetch::{SetFetch, SetFetchRequest, SetFetchValue};
-pub use requests::set::set_remove_elements::{SetRemoveElements, SetRemoveElementsRequest};
+pub use messages::set::set_add_elements::{SetAddElements, SetAddElementsRequest};
+pub use messages::set::set_fetch::{SetFetch, SetFetchRequest, SetFetchValue};
+pub use messages::set::set_remove_elements::{SetRemoveElements, SetRemoveElementsRequest};
 
-pub use requests::sorted_set::sorted_set_fetch_by_rank::{
+pub use messages::sorted_set::sorted_set_fetch_by_rank::{
     SortedSetFetchByRankRequest, SortedSetOrder,
 };
-pub use requests::sorted_set::sorted_set_fetch_by_score::SortedSetFetchByScoreRequest;
-pub use requests::sorted_set::sorted_set_fetch_response::{SortedSetElements, SortedSetFetch};
-pub use requests::sorted_set::sorted_set_get_rank::{SortedSetGetRank, SortedSetGetRankRequest};
-pub use requests::sorted_set::sorted_set_get_score::{SortedSetGetScore, SortedSetGetScoreRequest};
-pub use requests::sorted_set::sorted_set_length::{SortedSetLength, SortedSetLengthRequest};
-pub use requests::sorted_set::sorted_set_put_element::{
+pub use messages::sorted_set::sorted_set_fetch_by_score::SortedSetFetchByScoreRequest;
+pub use messages::sorted_set::sorted_set_fetch_response::{SortedSetElements, SortedSetFetch};
+pub use messages::sorted_set::sorted_set_get_rank::{SortedSetGetRank, SortedSetGetRankRequest};
+pub use messages::sorted_set::sorted_set_get_score::{SortedSetGetScore, SortedSetGetScoreRequest};
+pub use messages::sorted_set::sorted_set_length::{SortedSetLength, SortedSetLengthRequest};
+pub use messages::sorted_set::sorted_set_put_element::{
     SortedSetPutElement, SortedSetPutElementRequest,
 };
-pub use requests::sorted_set::sorted_set_put_elements::{
+pub use messages::sorted_set::sorted_set_put_elements::{
     IntoSortedSetElements, SortedSetElement, SortedSetPutElements, SortedSetPutElementsRequest,
 };
-pub use requests::sorted_set::sorted_set_remove_elements::{
+pub use messages::sorted_set::sorted_set_remove_elements::{
     SortedSetRemoveElements, SortedSetRemoveElementsRequest,
 };
 
-pub use requests::list::list_concatenate_back::{ListConcatenateBack, ListConcatenateBackRequest};
-pub use requests::list::list_concatenate_front::{
+pub use messages::list::list_concatenate_back::{ListConcatenateBack, ListConcatenateBackRequest};
+pub use messages::list::list_concatenate_front::{
     ListConcatenateFront, ListConcatenateFrontRequest,
 };
-pub use requests::list::list_fetch::{ListFetch, ListFetchRequest, ListFetchValue};
-pub use requests::list::list_length::{ListLength, ListLengthRequest};
-pub use requests::list::list_pop_back::{ListPopBack, ListPopBackRequest, ListPopBackValue};
-pub use requests::list::list_pop_front::{ListPopFront, ListPopFrontRequest, ListPopFrontValue};
-pub use requests::list::list_remove_value::{ListRemoveValue, ListRemoveValueRequest};
+pub use messages::list::list_fetch::{ListFetch, ListFetchRequest, ListFetchValue};
+pub use messages::list::list_length::{ListLength, ListLengthRequest};
+pub use messages::list::list_pop_back::{ListPopBack, ListPopBackRequest, ListPopBackValue};
+pub use messages::list::list_pop_front::{ListPopFront, ListPopFrontRequest, ListPopFrontValue};
+pub use messages::list::list_remove_value::{ListRemoveValue, ListRemoveValueRequest};
 
 // Similar re-exporting with config::configuration and config::configurations
 // so import paths can be simpmlified to "momento::cache::Configuration" and
